requirements.txt
openai==1.14.1
python-dotenv==1.0.1
qdrant-client==1.8.0
PyYAML==6.0.1
uvicorn==0.28.0
fastapi==0.110.0
Jinja2==3.1.3
python-multipart==0.0.9
websockets==12.0